% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_pdf.R
\name{check_libreoffice_export}
\alias{check_libreoffice_export}
\title{Check if PDF export is functional}
\usage{
check_libreoffice_export(UserInstallation = NULL)
}
\arguments{
\item{UserInstallation}{use this value to set a non-default user profile path
for "LibreOffice". If not provided a temporary dir is created. It makes possibles
to use more than a single session of "LibreOffice."}
}
\value{
a single logical value.
}
\description{
Test if 'LibreOffice' can export to PDF.
An attempt to export to PDF is made to confirm that
the PDF export is functional.
}
\examples{
library(locatexec)
if(exec_available("libreoffice")){
  \donttest{check_libreoffice_export()}
}
}
