% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx2pdf.R
\name{docx2pdf_available}
\alias{docx2pdf_available}
\title{Is 'docx2pdf' available}
\usage{
docx2pdf_available(error = FALSE)
}
\arguments{
\item{error}{Whether to signal an error if 'docx2pdf' is not found}
}
\value{
a single logical value.
}
\description{
Checks if 'docx2pdf' is available within a given 'python' distribution.
}
\examples{
library(locatexec)

if(exec_available("python") &&
   exec_version("python") > numeric_version("3")) {
  docx2pdf_available()
}
}
\seealso{
Other tools for docx2pdf: 
\code{\link{docx2pdf_install}()},
\code{\link{docx2pdf_uninstall}()},
\code{\link{docx2pdf}()}
}
\concept{tools for docx2pdf}
