% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_miniature.R
\name{msoffice_available}
\alias{msoffice_available}
\title{Check if 'Microsoft Office' is available}
\usage{
msoffice_available()
}
\value{
a single logical value.
}
\description{
The function test if 'Microsoft Office' is available.
}
\examples{
msoffice_available()
}
