% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect_snapshot_doc.R
\name{skip_if_not_snapshot_png}
\alias{skip_if_not_snapshot_png}
\title{Skip tests when PNG snapshot dependencies are unavailable}
\usage{
skip_if_not_snapshot_png()
}
\value{
Invisibly returns NULL or skips the test.
}
\description{
Skips a testthat test if ragg or gdtools are not installed.
Also registers Liberation Sans via \code{\link[gdtools:register_liberationsans]{gdtools::register_liberationsans()}}.
}
