% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_pdf.R
\name{to_pdf}
\alias{to_pdf}
\title{Convert documents to pdf}
\usage{
to_pdf(input, output = gsub("\\\\.[[:alnum:]]+$", ".pdf", input))
}
\arguments{
\item{input, output}{file input and optional file output.}
}
\description{
Convert documents to pdf using Libre Office. It
supports very well "Microsoft PowerPoint" to PDF. "Microsoft Word"
can also be converted but some Word features are not supported
such as sections.
}
\examples{
library(locatexec)
if (exec_available("libreoffice")) {
  file <- tempfile(fileext = ".pptx")
  file.copy(
    system.file(package = "doconv",
      "doc-examples/first_example.pptx"),
    file)

  out <- to_pdf(input = file)
}
}
