% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx2pdf.R
\name{xlsx2pdf}
\alias{xlsx2pdf}
\title{Convert xlsx to pdf}
\usage{
xlsx2pdf(input, output = gsub("\\\\.(xlsx|xlsm)$", ".pdf", input))
}
\arguments{
\item{input, output}{file input and optional file output. If output
file is not provided, the value will be the value of input file with
extension 'pdf'.}
}
\value{
the name of the produced pdf (the same value as \code{output})
}
\description{
Convert xlsx to pdf directly using "Microsoft Excel".

This function will not work if "Microsoft Excel" is not available
on your machine.

The calls to "Microsoft Excel" are made differently depending on the
operating system:
\itemize{
\item On "Windows", a "PowerShell" script using COM
technology is used to control "Microsoft Excel".
\item On macOS, an "AppleScript" script is used to control "Microsoft Excel".
}
}
\section{Windows authorizations}{

If your execution policy is set to "RemoteSigned", 'doconv' will
not be able to run powershell script. Set it to "Unrestricted" and it
should work. If you are in a managed and administrated environment,
you may not be able to use 'doconv' because of execution policies.
}

\section{Macos manual authorizations}{

On macOS the call is happening into a working
directory managed with function \code{\link[=working_directory]{working_directory()}}.

Manual interventions are necessary to authorize 'Excel'
application to write in a single directory: the working directory.
These permissions must be set manually, this is required by the macOS security
policy.
}

\examples{
library(locatexec)
if (exec_available("excel")) {
  xlsx2pdf(
    input = system.file(
      package = "doconv",
      "doc-examples/example.xlsx"
    ),
    output = tempfile(fileext = ".pdf")
  )
}
}
