% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx2pdf.R
\name{docx2pdf_uninstall}
\alias{docx2pdf_uninstall}
\title{Uninstall 'docx2pdf'}
\usage{
docx2pdf_uninstall()
}
\value{
a single logical value, FALSE if the operation failed, TRUE otherwise.
}
\description{
Removes 'docx2pdf' within a given 'python' distribution.
}
\examples{
library(locatexec)

if(exec_available("python") &&
   exec_version("python") > numeric_version("3") &&
   docx2pdf_available()) {
  \donttest{
  # this will uninstall and install 'docx2pdf' (i.e. `pip install docx2pdf`)
  # in your python environnement. It can run during 10s and
  # require user to have write permission in python environnement.
  docx2pdf_uninstall()
  docx2pdf_install()
  }
}
}
\seealso{
Other tools for docx2pdf: 
\code{\link{docx2pdf_available}()},
\code{\link{docx2pdf_install}()},
\code{\link{docx2pdf}()}
}
\concept{tools for docx2pdf}
