% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_miniature.R
\name{to_miniature}
\alias{to_miniature}
\title{Thumbnail of a document}
\usage{
to_miniature(
  filename,
  row = NULL,
  width = NULL,
  border_color = "#ccc",
  border_geometry = "2x2",
  fileout = NULL,
  use_docx2pdf = FALSE
)
}
\arguments{
\item{filename}{input file}

\item{row}{row index for every pages}

\item{width}{width of a single image}

\item{border_color}{border color}

\item{border_geometry}{border geometry to be added around images}

\item{fileout}{if not NULL, result is saved in a png file whose filename
is defined by this argument.}

\item{use_docx2pdf}{if TRUE (and if 'Microsoft Word' executable
can be found as well as 'docx2pdf'), docx2pdf will be used to
convert 'Word' documents to PDF. This makes it possible to have a
PDF identical to the 'Word' display whereas with 'LibreOffice', this
is not always the case.}
}
\value{
a magick image object as returned by \code{\link[=image_read]{image_read()}}.
}
\description{
Convert a file to an image (magick image)
where pages are arranged in a layout. Result can be saved
to a png file.
}
\examples{
library(locatexec)
docx_file <- system.file(
  package = "doconv",
  "doc-examples/bookdown.docx"
)
if(exec_available("python") && docx2pdf_available())
  to_miniature(docx_file)

pptx_file <- system.file(
  package = "doconv",
  "doc-examples/first_example.pptx"
)
if(exec_available("libreoffice"))
  to_miniature(pptx_file)
}
