% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_pdf.R
\name{to_pdf}
\alias{to_pdf}
\title{Convert documents to pdf}
\usage{
to_pdf(
  input,
  output = gsub("\\\\.[[:alnum:]]+$", ".pdf", input),
  use_docx2pdf = FALSE
)
}
\arguments{
\item{input, output}{file input and optional file output.}

\item{use_docx2pdf}{if TRUE (and if 'Microsoft Word' executable
can be found as well as 'docx2pdf'), docx2pdf will be used to
convert 'Word' documents to PDF. This makes it possible to have a
PDF identical to the 'Word' display whereas with 'LibreOffice', this
is not always the case.}
}
\description{
Convert documents to pdf using Libre Office. It
supports very well "Microsoft PowerPoint" to PDF. "Microsoft Word"
can also be converted but some Word features are not supported
such as sections.
}
\examples{
library(locatexec)
if (exec_available("libreoffice")) {
  file <- tempfile(fileext = ".pptx")
  file.copy(
    system.file(package = "doconv",
      "doc-examples/first_example.pptx"),
    file)

  out <- to_pdf(input = file)
}
}
